Rename old i18n implementation folder to `i18n-legacy`
