+++ both/api/categories/server/transifex-methods.js
@@ -1,4 +1,4 @@
 import { Categories } from '/both/api/categories/categories';
-import { makeCollectionTranslatable } from '/server/i18n/translatable-collection';
+import { makeCollectionTranslatable } from '/server/i18n-legacy/translatable-collection';
 
 makeCollectionTranslatable(Categories, '_id');
