+++ both/api/sources/server/transifex-methods.js
@@ -1,4 +1,4 @@
 import { Sources } from '/both/api/sources/sources';
-import { makeCollectionTranslatable } from '/server/i18n/translatable-collection';
+import { makeCollectionTranslatable } from '/server/i18n-legacy/translatable-collection';
 
 makeCollectionTranslatable(Sources, 'additionalAccessibilityInformation');
