+++ server/i18n-legacy/collection-schema.js
