+++ server/i18n-legacy/sync-with-transifex.js
