+++ server/i18n-legacy/sync.js
