+++ server/i18n-legacy/translatable-collection.js
